// petstore/servers/node/tsp-output/@typespec/http-server-javascript/models/all/pet-store.ts
export interface Pets<Context = unknown> {
  /**
   * Gets an instance of the resource.
   */
  get(ctx: Context, petId: number): Promise<Pet | PetStoreError>;

  /**
   * Updates an existing instance of the resource.
   */
  update(
    ctx: Context,
    petId: number,
    properties: PetUpdate,
  ): Promise<Pet | PetStoreError>;

  /**
   * Deletes an existing instance of the resource.
   */
  delete(
    ctx: Context,
    petId: number,
  ): Promise<ResourceDeletedResponse | PetStoreError>;

  /**
   * Creates a new instance of the resource.
   */
  create(
    ctx: Context,
    resource: PetCreate,
  ): Promise<Pet | PetResourceCreatedResponse | PetStoreError>;

  /**
   * Lists all instances of the resource.
   */
  list(ctx: Context): Promise<PetCollectionWithNextLink | PetStoreError>;
}
// ...
export interface PetStoreError {
  code: number;

  message: string;
}

export interface Pet {
  id: number;

  name: string;

  tag?: string;

  age: number;

  ownerId: bigint;
}